Why have two casts
when you can have one

+++ src/player.rs
@@ -423,19 +423,19 @@ impl Player {
     }
 
     pub fn set_level(&mut self, level: i16) {
-        self.combat_stats.level = clamp(level as u32, 1, PC_LEVEL_MAX) as i16;
+        self.combat_stats.level = clamp(level, 1, PC_LEVEL_MAX as i16);
     }
 
     pub fn set_fusion_matter(&mut self, fusion_matter: i32) {
-        self.fusion_matter = clamp(fusion_matter as u32, 0, PC_FUSIONMATTER_MAX) as i32;
+        self.fusion_matter = clamp(fusion_matter, 0, PC_FUSIONMATTER_MAX as i32);
     }
 
     pub fn set_weapon_boosts(&mut self, weapon_boosts: i32) {
-        self.weapon_boosts = clamp(weapon_boosts as u32, 0, PC_BATTERY_MAX) as i32;
+        self.weapon_boosts = clamp(weapon_boosts, 0, PC_BATTERY_MAX as i32);
     }
 
     pub fn set_nano_potions(&mut self, nano_potions: i32) {
-        self.nano_potions = clamp(nano_potions as u32, 0, PC_BATTERY_MAX) as i32;
+        self.nano_potions = clamp(nano_potions, 0, PC_BATTERY_MAX as i32);
     }
 
     pub fn set_god_mode(&mut self, god_mode: bool) {
